WIP: the automatic sampling should work this way:
When the sensor starts, discard measurements for a pre-determined amount of time.
Then, for a chosen interval, aggregate measuments and calculate mean values.
Go on reporting until the sensor stops emitting measurements for set duration.
Wait for the next start, then repeat above sequence.
Keep on this style of aggregated reporting until the handler is reset.